+++ src/source/math/group_c/set_c.hpp
@@ -7,7 +7,7 @@
 
 namespace dla
 {
-    class set_c : public set<item_c, uint16_t> {
+    class set_c : public set<item_c, uint32_t> {
     private:
         static const int PAYLOAD_SIZE = 15;
         static const int ITEM_SIZE = 16;
